Fix simulated region schemas from review
