+++ shared/src/simulation/behaviors/simulation-behavior.ts
@@ -1,12 +1,12 @@
 import type { WritableDraft } from 'immer';
-import { string, z } from 'zod';
+import { z } from 'zod';
 import type { ExerciseState } from '../../state.js';
 import { uuidSchema } from '../../utils/uuid.js';
 import type { SimulatedRegion } from '../../models/simulated-region.js';
 import type { ExerciseSimulationEvent } from '../events/exercise-simulation-event.js';
 
 export const simulationBehaviorStateSchema = z.strictObject({
-    type: z.literal(`${string}Behavior`),
+    type: z.templateLiteral([z.string(), `Behavior`]),
     id: uuidSchema,
 });
