+++ shared/src/simulation/behaviors/utils.ts
@@ -1,3 +1,4 @@
+import { z } from 'zod';
 import { newMaterialCountRadiogram } from '../../models/radiogram/material-count-radiogram.js';
 import { newPatientCountRadiogram } from '../../models/radiogram/patient-count-radiogram.js';
 import { newPersonnelCountRadiogram } from '../../models/radiogram/personnel-count-radiogram.js';
@@ -10,8 +11,42 @@ import { newVehicleOccupationsRadiogram } from '../../models/radiogram/vehicle-o
 import type { UUID } from '../../utils/uuid.js';
 import type { ExerciseRadiogramStatus } from '../../models/radiogram/status/exercise-radiogram-status.js';
 import type { ExerciseRadiogram } from '../../models/radiogram/exercise-radiogram.js';
-import type { ReportableInformation } from './reportable-information.js';
 import type { ExerciseSimulationBehaviorType } from './exercise-simulation-behavior.js';
+
+export const reportableInformationAllowedValues = [
+    'materialCount',
+    'patientCount',
+    'personnelCount',
+    'requiredResources',
+    'singleRegionTransferCounts',
+    'transferConnections',
+    'transportManagementTransferCounts',
+    'treatmentStatus',
+    'vehicleCount',
+    'vehicleOccupations',
+] as const;
+
+export const reportableInformationSchema = z.literal(
+    reportableInformationAllowedValues
+);
+export type ReportableInformation = z.infer<typeof reportableInformationSchema>;
+
+export const reportableInformationTypeToGermanNameDictionary: {
+    [Key in ReportableInformation]: string;
+} = {
+    materialCount: 'Anzahl an Material',
+    patientCount: 'Anzahl an Patienten',
+    personnelCount: 'Anzahl an Rettungskräften',
+    requiredResources: 'Aktuell benötigte Fahrzeuge',
+    singleRegionTransferCounts:
+        'Anzahl aus diesem Bereich in Krankenhäuser abtransportierter Patienten',
+    transferConnections: 'Transferverbindungen',
+    transportManagementTransferCounts:
+        'Anzahl unter dieser Transportorganisation in Krankenhäuser abtransportierter Patienten',
+    treatmentStatus: 'Behandlungsstatus',
+    vehicleCount: 'Anzahl an Fahrzeugen',
+    vehicleOccupations: 'Nutzung der Fahrzeuge',
+};
 
 export const createRadiogramMap: {
     [Key in ReportableInformation]: (
